+++ Scr/World/Chunk/Mesh.h
@@ -7,18 +7,20 @@ namespace cubecraft {
 	class Mesh {
 	public:
 		Mesh() = default;
-		Mesh(std::vector<GLfloat> _vertices, std::vector<GLuint> _vIndices,
-			std::vector<GLuint> _tIndices) {
+		Mesh(
+			std::vector<GLfloat>	_vertices,
+			std::vector<GLuint>		_vIndices,
+			std::vector<GLfloat>	_tIndices) {
 			vertices = _vertices;
 			vIndices = _vIndices;
 			tIndices = _tIndices;
 		}
 		std::vector<GLfloat> getVertices();
 		std::vector<GLuint> getVer_Indices();
-		std::vector<GLuint> getTex_Indices();
+		std::vector<GLfloat> getTex_Indices();
 	private:
 		std::vector<GLfloat> vertices; // 顶点
 		std::vector<GLuint> vIndices; // 顶点索引
-		std::vector<GLuint> tIndices; // 纹理索引
+		std::vector<GLfloat> tIndices; // 纹理索引
 	};
 }
